test: route xarray-only tests to the xarray job, off the core suite (#638)
Close the residual 'core test that silently skips without xarray' cases and the
pure-wheel coverage gap for the xarray-consumer tests.

- cog facade: move the xr.DataArray-input tests (TestDataArrayToDataset +
  the DataArray dispatch) out of the core test_facade.py into a new
  xarray-marked test_facade_xarray.py, so the bare-wheel core job no longer
  collects-and-skips them.
- DatasetCollection.to_netcdf: mark the module xarray (to_netcdf needs xarray)
  and split the missing-xarray branch into test_to_netcdf_missing_xarray.py,
  which stays core so it runs in the extras-free suite.
- kerchunk consumers (test_io_e2e, test_kerchunk_native,
  test_netcdf_lazy_pipelines): replace the hand-rolled requires_xarray skipif
  with @pytest.mark.xarray so they actually run in the xarray CI cell (which
  installs lazy + xarray) instead of skipping in the lazy/netcdf_lazy cells and
  never being selected by -m xarray; the guarded import keeps collection safe
  without xarray.
- pure-wheel-test.yml: correct the stale comment — core is an explicit
  pytestmark, not auto-applied; the conftest hook only auto-skips extras tests.
